Add pagination, filtering, sorting for geolocations

+++ src/services/GeolocationService.ts
@@ -1,5 +1,5 @@
 import { Model } from 'mongoose';
-import { BadRequestException, HttpException, Injectable, Logger } from '@nestjs/common';
+import { BadRequestException, HttpException, HttpStatus, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { GeolocationDocument } from 'src/schemas/Geolocation';
 import { ConfigService } from '@nestjs/config';
@@ -7,6 +7,7 @@ import { AxiosFactory } from '../providers/AxiosFactory';
 import { AxiosInstance } from 'axios';
 import { DeleteResult } from 'mongodb';
 import { IndexGeolocationDto } from 'src/dto/geolocation/IndexGeolocationDto';
+import { IIndexGeolocations } from 'src/interfaces/geolocations/IIndexGeolocations';
 
 @Injectable()
 export class GeolocationService {
@@ -37,21 +38,60 @@ export class GeolocationService {
 
             Logger.error(ipStackError);
 
-            throw new HttpException('Geolocation service error', 424);
+            throw new HttpException('Geolocation service error', HttpStatus.FAILED_DEPENDENCY);
         }
 
         return this.geolocationModel.create(data);
     }
 
-    getAll({ skip, limit }: IndexGeolocationDto): Promise<GeolocationDocument[]> {
-        return this.geolocationModel.find().skip(skip).limit(limit).exec();
+    async getAll({ skip, limit, sort, order, searchTerm }: IndexGeolocationDto): Promise<IIndexGeolocations> {
+        const geolocations = await this.geolocationModel
+            .find({
+                $or: [
+                    { ip: { $regex: searchTerm } },
+                    { continent_name: { $regex: searchTerm } },
+                    { country_name: { $regex: searchTerm } },
+                    { city: { $regex: searchTerm } }
+                ]
+            })
+            .skip(skip)
+            .limit(limit)
+            .sort([[order, sort]])
+            .exec();
+
+        const totalGeolocationsCount = await this.geolocationModel
+            .find({
+                $or: [
+                    { ip: { $regex: searchTerm } },
+                    { continent_name: { $regex: searchTerm } },
+                    { country_name: { $regex: searchTerm } },
+                    { city: { $regex: searchTerm } }
+                ]
+            })
+            .countDocuments();
+
+        return {
+            geolocations,
+            totalGeolocationsCount
+        };
     }
 
-    getAllByIp({ ip, skip, limit }: IndexGeolocationDto): Promise<GeolocationDocument[]> {
-        return this.geolocationModel.find({ ip }).skip(skip).limit(limit).exec();
+    getAllByIp({ ip, skip, limit, sort, order }: IndexGeolocationDto): Promise<GeolocationDocument[]> {
+        return this.geolocationModel
+            .find({ ip })
+            .skip(skip)
+            .limit(limit)
+            .sort([[order, sort]])
+            .exec();
     }
 
-    delete(id: string): Promise<DeleteResult> {
+    async delete(id: string): Promise<DeleteResult> {
+        const geolocationToDelete = await this.geolocationModel.findById(id);
+
+        if (!geolocationToDelete) {
+            throw new NotFoundException('Resource does not exist');
+        }
+
         return this.geolocationModel.deleteOne({ _id: id }).exec();
     }
 }
